Authorization 1

+++ app/Http/Controllers/Dashboard/AdminController.php
@@ -4,35 +4,48 @@ namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
 use App\Models\Admin;
+use App\Models\Role;
+use App\Models\RoleUser;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
+use Inertia\Inertia;
+use PhpParser\Node\Expr\Throw_;
+use Throwable;
+use function PHPUnit\Framework\throwException;
 
 class AdminController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
-        $admins = Admin::paginate();
-        return view('dashboard.admins.index', compact('admins'));
+        Gate::authorize('view admins');
+        $admins = Admin::with('roles')->where('super_admin' , false)->paginate();
+        return Inertia::render(
+            'dashboard/admins/admins.index',
+            ['admins' => $admins]
+        );
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
-        $admin = new Admin();
-        return view('dashboard.admins.create', compact('admin'));
+        Gate::authorize('create admins');
 
+        $admin = new Admin();
+
+        return Inertia::render(
+            'dashboard/admins/admins.create',
+            [
+                'admin' => $admin,
+                'roles' => Role::all(),
+            ]
+        );
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
+        Gate::authorize('create admins');
+
         $request->validate([
             'name' => 'required|string|max:255',
             'username' => 'required|string|max:255',
@@ -40,36 +53,54 @@ class AdminController extends Controller
             'password' => 'required',
             'phone_number' => 'required|numeric',
             'status' => 'required|in:active,inactive',
+            'roles' => 'nullable|array',
         ]);
 
-        $data = $request->except('password');
-        $data['password'] = Hash::make($request->post('password'));
 
-        $admin = Admin::create($data);
-        return redirect()->route('dashboard.admins.index')->with('success', 'Admin Added Successfully');
+        DB::beginTransaction();
+        try {
+            $data = $request->except(['password', 'roles']);
+            $data['password'] = Hash::make($request->post('password'));
+            $admin = Admin::create($data);
+
+            if ($request->post('roles')) {
+                foreach ($request->post('roles') as $role) {
+                    $admin->roles()->attach($role['id']);
+                }
+            }
+            DB::commit();
+        } catch (Throwable $e) {
+            DB::rollBack();
+            throw $e;
+        }
+
+
+        return redirect()->route('dashboard.admins.index')->with('message', 'Admin Added Successfully');
     }
 
-    /**
-     * Display the specified resource.
-     */
     public function show(string $id)
     {
-        //
+
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
     public function edit(Admin $admin)
     {
-        return view('dashboard.admins.edit', compact('admin'));
+        Gate::authorize('update admins');
+
+        return Inertia::render(
+            'dashboard/admins/admins.edit',
+            [
+                'admin' => $admin->load('roles'),
+                'roles' => Role::all(),
+            ]
+        );
+        ;
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, Admin $admin)
     {
+        Gate::authorize('update admins');
+
         $request->validate([
             'name' => 'required|string|max:255',
             'username' => 'required|string|max:255',
@@ -77,15 +108,27 @@ class AdminController extends Controller
             'password' => 'required',
             'phone_number' => 'required|numeric',
             'status' => 'required|in:active,inactive',
+            'roles' => 'nullable|array',
         ]);
 
-        $data = $request->except('password');
+        $data = $request->except(['password', 'roles']);
         $data['password'] = Hash::make($request->post('password'));
 
 
         $admin->update($data);
+
+        if ($request->post('roles')) {
+            $ids = [];
+            foreach ($request->post('roles') as $role) {
+                $ids[] = $role['id'];
+            }
+            $admin->roles()->sync($ids);
+        } else {
+            $admin->roles()->detach();
+        }
+
         return redirect()->route('dashboard.admins.index')
-            ->with('success', 'Admin Updated Successfully');
+            ->with('message', 'Admin Updated Successfully');
 
     }
 
@@ -94,9 +137,7 @@ class AdminController extends Controller
      */
     public function destroy(Admin $admin)
     {
+        Gate::authorize('delete admins');
         $admin->delete();
-        return redirect()->route('dashboard.admins.index')
-            ->with('success', 'Admin Deleted Successfully');
-
     }
 }
