+++ app/Http/Controllers/Dashboard/RoleController.php
@@ -3,23 +3,23 @@
 namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
-use App\Models\Category;
 use App\Models\Role;
 use App\Models\RoleAbility;
-use App\Models\RoleUser;
+use Exception;
+use Illuminate\Database\QueryException;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
-use phpDocumentor\Reflection\Types\Object_;
+use function PHPUnit\Framework\throwException;
 
 class RoleController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
+        Gate::authorize('view roles');
+
         $roles = Role::paginate();
 
         return Inertia::render(
@@ -28,34 +28,26 @@ class RoleController extends Controller
         );
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
+
     public function create()
     {
+        Gate::authorize('create roles');
+
         $role = new Role();
-        $abilities = config('abilities');
-
-
-        $newArray = [];
-        foreach ($abilities as $key => $value) {
-            $newArray[] = (Object) ['ability' => $key, 'value' => $value];
-        }
 
         return Inertia::render(
             'dashboard/roles/roles.create',
             [
                 'role' => $role,
-                'abilities' => $abilities,
+                'allAbilities' => Config::get('abilities'),
             ],
         );
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
+        Gate::authorize('create roles');
+
         $request->validate([
             'name' => 'required|string|max:255',
             'abilities' => 'required|array',
@@ -67,80 +59,60 @@ class RoleController extends Controller
             ]);
 
             $abilities = $request->post('abilities');
-
-            foreach ($abilities as $ability_name => $value) {
-                RoleAbility::create([
-                    'role_id' => $role->id,
-                    'ability' => $ability_name,
-                    'type' => $value,
-                ]);
-            }
-
-
-            // RoleUser::create([
-            //   'authorizable_type' => get_class(Auth::user()),
-            //   'authorizable_id' => Auth::id(),
-            //   'role_id'=>$role->id ,
-            // ]);
+            $this->storeAbilities($abilities, $role->id);
 
             DB::commit();
-        } catch (\Exception $e) {
+        } catch (QueryException $e) {
             DB::rollBack();
-            throw $e;
+            report(
+                $e
+            );
         }
 
 
         return redirect()
             ->route('dashboard.roles.index')
             ->with('message', 'Role Added Successfully');
-
-
     }
 
 
     public function show(string $id)
     {
-        //
     }
 
     public function edit(Role $role)
     {
-        $abilities = config('abilities');
+        Gate::authorize('update roles');
+
         return Inertia::render(
             'dashboard/roles/roles.edit',
             [
-                'role' => $role,
-                'abilities' => $abilities,
+                'role' => $role->load('abilities'),
+                'allAbilities' => config('abilities'),
             ],
         );
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
+
     public function update(Request $request, Role $role)
     {
+        Gate::authorize('update roles');
+
         $request->validate([
             'name' => 'required|string|max:255',
             'abilities' => 'required|array',
         ]);
+
         DB::beginTransaction();
         try {
             $role->update([
                 'name' => $request->post('name'),
             ]);
 
-            $abilities = $request->post('abilities');
-            foreach ($abilities as $ability_name => $value) {
-                RoleAbility::updateOrCreate([
-                    'role_id' => $role->id,
-                    'ability' => $ability_name,
-                ], [
-                    'type' => $value,
-                ]);
-            }
+            $this->updateAbilities($request->post('abilities'), $role->id);
+
             DB::commit();
-        } catch (\Exception $e) {
+        } catch (QueryException $e) {
             DB::rollBack();
             throw $e;
         }
@@ -153,6 +125,30 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
+        Gate::authorize('delete roles');
+
         $role->delete();
+    }
+
+    public function storeAbilities($abilities, $role_id)
+    {
+        foreach ($abilities as $ability) {
+            RoleAbility::create([
+                'role_id' => $role_id,
+                'ability' => $ability['name'],
+                'type' => $ability['type'],
+            ]);
+        }
+    }
+    public function updateAbilities($abilities, $role_id)
+    {
+        foreach ($abilities as $ability) {
+            RoleAbility::updateOrCreate([
+                'role_id' => $role_id,
+                'ability' => $ability['ability'],
+            ], [
+                'type' => $ability['type'],
+            ]);
+        }
     }
 }
